images-12-generation/pyproject.toml
[project]
name = "xampler-images-12-generation"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "cfboundary @ git+https://github.com/adewale/cfboundary@v0.1.10",
    "pillow>=11",
]

[dependency-groups]
dev = [
    "workers-py",
    "workers-runtime-sdk",
]
